Refactor JERI code
